Provide a logging API for netstandard2.0
The logging API is a no-op in netstandard2.0 in order for multi-targeted and .NET Standard consumers to still be able to use it. These assemblies will still get the real deal when used in a .NET 6 app.

This splits the library into 2 projects in order to avoid having to write #if NETCOREAPP literally everywhere. It is also less error-prone since the base API is isolated.